Update Zn_NH3_solution
Updated Zn_NH3_olution to include a dictionary of constants from the one article. Also, altered the sequence for the concentration_array to be more systematic. Changed that in the plotting.
The script also contains equilibrium reactions which was present in the paper, but they do not affect the already made script since they are not included in the conservation equations